etl/requirements.txt
psycopg2-binary==2.9.3
pydantic
redis==2.10.6
elasticsearch-dsl==7.4.0
backoff
